refactor: mv Queue class to streamr-test-utils
