+++ packages/broker/test/integration/multiple-publisher-plugins.test.ts
@@ -3,8 +3,9 @@ import WebSocket from 'ws'
 import fetch from 'node-fetch'
 import { StreamPermission } from 'streamr-client'
 import { Tracker } from '@streamr/network-tracker'
+import { Queue } from 'streamr-test-utils'
 import { Broker } from '../../src/broker'
-import { startBroker, createClient, createTestStream, fetchPrivateKeyWithGas, startTestTracker, Queue } from '../utils'
+import { startBroker, createClient, createTestStream, fetchPrivateKeyWithGas, startTestTracker } from '../utils'
 import { fastPrivateKey, waitForCondition, waitForEvent } from 'streamr-test-utils'
 import { range, sample } from 'lodash'
