+++ packages/broker/test/integration/plugins/mqtt/Bridge.test.ts
@@ -1,8 +1,9 @@
 import { Stream, StreamrClient } from 'streamr-client'
 import { Tracker } from '@streamr/network-tracker'
 import mqtt from 'async-mqtt'
+import { Queue } from 'streamr-test-utils'
 import { Broker } from '../../../../src/broker'
-import { createClient, startBroker, createTestStream, Queue, fetchPrivateKeyWithGas, startTestTracker } from '../../../utils'
+import { createClient, startBroker, createTestStream, fetchPrivateKeyWithGas, startTestTracker } from '../../../utils'
 import { wait } from 'streamr-test-utils'
 import { Wallet } from '@ethersproject/wallet'
